refactor: move test validation into a different class
